requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
idna==2.10
lxml==4.6.2
multidict==5.1.0
pkg-resources==0.0.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
soupsieve==2.2
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
